refactor: move Propagate class to new propagate.py
