+++ gmatpyplus/command/__init__.py
@@ -1,5 +1,6 @@
 from .branch_command import *
 from .command import *
 from .gmat_command import GmatCommand
+from .propagate import Propagate
 from .solver_branch_command import *
 from .solver_sequence_command import *
